Working... ish

+++ src/feature/map/utilities/U.java
@@ -53,7 +53,7 @@ public class U {
      * @return 
      */
     public static int l (double d) {
-        double D = d * 255;
+        int D = (int) (d * 255);
         
         D = (D < 0)? 0 : D;
         D = (D > 255)? 255 : D;
